Move MT broker lifecycle and handle only one broker class
