+++ pkg/apis/eventing/v1beta1/broker_lifecycle_mt.go
@@ -1,0 +1,62 @@
+/*
+ * Copyright 2020 The Knative Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package v1beta1
+
+import (
+	corev1 "k8s.io/api/core/v1"
+
+	"knative.dev/eventing/pkg/apis/duck"
+	duckv1beta1 "knative.dev/eventing/pkg/apis/duck/v1beta1"
+)
+
+func (bs *BrokerStatus) MarkIngressFailed(reason, format string, args ...interface{}) {
+	bs.GetConditionSet().Manage(bs).MarkFalse(BrokerConditionIngress, reason, format, args...)
+}
+
+func (bs *BrokerStatus) PropagateIngressAvailability(ep *corev1.Endpoints) {
+	if duck.EndpointsAreAvailable(ep) {
+		bs.GetConditionSet().Manage(bs).MarkTrue(BrokerConditionIngress)
+	} else {
+		bs.MarkIngressFailed("EndpointsUnavailable", "Endpoints %q are unavailable.", ep.Name)
+	}
+}
+
+func (bs *BrokerStatus) MarkTriggerChannelFailed(reason, format string, args ...interface{}) {
+	bs.GetConditionSet().Manage(bs).MarkFalse(BrokerConditionTriggerChannel, reason, format, args...)
+}
+
+func (bs *BrokerStatus) PropagateTriggerChannelReadiness(cs *duckv1beta1.ChannelableStatus) {
+	// TODO: Once you can get a Ready status from Channelable in a generic way, use it here...
+	address := cs.AddressStatus.Address
+	if address != nil {
+		bs.GetConditionSet().Manage(bs).MarkTrue(BrokerConditionTriggerChannel)
+	} else {
+		bs.MarkTriggerChannelFailed("ChannelNotReady", "trigger Channel is not ready: not addressable")
+	}
+}
+
+func (bs *BrokerStatus) MarkFilterFailed(reason, format string, args ...interface{}) {
+	bs.GetConditionSet().Manage(bs).MarkFalse(BrokerConditionFilter, reason, format, args...)
+}
+
+func (bs *BrokerStatus) PropagateFilterAvailability(ep *corev1.Endpoints) {
+	if duck.EndpointsAreAvailable(ep) {
+		bs.GetConditionSet().Manage(bs).MarkTrue(BrokerConditionFilter)
+	} else {
+		bs.MarkFilterFailed("EndpointsUnavailable", "Endpoints %q are unavailable.", ep.Name)
+	}
+}
